Creating formating for playing embed.
